Moving main page
