+++ web-group-buying-store/src/pages/MainPage.css
